Correctly implement PHP Coveralls in Travis
This commit fixes the currently broken Coveralls + Travis integration
by changing the way Travis pulled in the Coveralls dependency. In prior
my last package, ancarda/security-headers, the Travis script ran
composer require, causing a 1.x version of PHP Coveralls to be
installed, however now that 2.0 is out, and the path to the binary has
changed, Travis couldn't run PHP Coveralls as the file wasn't found.

This was change was caused by the following commit:

    https://github.com/php-coveralls/php-coveralls/commit/3eaf7eb689

Lesson Learned: You should check *everything* into composer.json, even
if said dependency is just for your Continual Integration system. That
way, when there are changes to those packages, you can better control
the build environment. Additionally, it allows you to take SemVer more
seriously, as that would have helped if I was pinned to 1.x of this
package.